Moved academics components to where it belongs
